erro na pagina search e paginação

+++ src/components/deleteModal/index.tsx
@@ -1,37 +1,43 @@
-'use client'
+// DeleteModal.tsx
+'use client';
 
 import { deleteProduct } from "../../../actions/home/actions";
+import { useRouter } from 'next/navigation'; // Importa o hook de navegação
 
-type ModalProps={
+type ModalProps = {
     isOpen: boolean;
     closeModal: () => void;
-    deleteId: number | undefined; //Produto que vai ser deletado
+    deleteId: number | undefined;
 };
-//Modal de Deletar
-export default function DeleteModal({ isOpen, closeModal, deleteId}: ModalProps){
+
+export default function DeleteModal({ isOpen, closeModal, deleteId }: ModalProps) {
+    const router = useRouter(); // Obtém o router para navegação
+
     const handleDelete = async () => {
-        await deleteProduct(deleteId);
-        alert("Produto deletado com sucesso");
-        closeModal();
+        if (deleteId) {
+            await deleteProduct(deleteId); // Chama a função de deletar
+            closeModal(); // Fecha o modal
+            router.push('/gerenciamento'); // Redireciona para a página de gerenciamento
+        }
     };
+
     if (!isOpen) {
-            return null;
+        return null;
     }
-    return(
+
+    return (
         <div className="fixed inset-0 flex items-center justify-center bg-black bg-opacity-50 z-50">
-        <div className="w-2/6 p-5 h-fit flex flex-col bg-rosaB rounded-md fixed">
-            <h1 className="text-rosaMarrom pb-3 font-bold text-xl">Deletar</h1>
-            
-            <div className="flex flex-col justify-center items-center px-3">
-                <span className="text-rosaMarrom text-center py-2 font-bold">Tem certeza que deseja deletar esse item?</span>
-                <span className="text-rosaMarrom text-center py-2 font-bold">Esta ação não poderá ser desfeita.</span>
+            <div className="w-2/6 p-5 h-fit flex flex-col bg-rosaB rounded-md">
+                <h1 className="text-rosaMarrom pb-3 font-bold text-xl">Deletar</h1>
+                <div className="flex flex-col justify-center items-center px-3">
+                    <span className="text-rosaMarrom text-center py-2 font-bold">Tem certeza que deseja deletar esse item?</span>
+                    <span className="text-rosaMarrom text-center py-2 font-bold">Esta ação não poderá ser desfeita.</span>
+                </div>
+                <div className="pt-9 flex gap-5 justify-end items-end">
+                    <button className="bg-rosaMarrom text-white px-3 py-1 rounded-full" onClick={handleDelete}>Deletar</button>
+                    <button className="bg-rosaMarrom text-white px-3 py-1 rounded-full" onClick={closeModal}>Cancelar</button>
+                </div>
             </div>
-            
-            <div className="pt-9 flex gap-5 justify-end items-end">
-                <button className="bg-rosaMarrom text-white px-3 py-1 rounded-full" onClick={handleDelete}>Deletar</button>
-                <button className="bg-rosaMarrom text-white px-3 py-1 rounded-full" onClick={closeModal}>Cancelar</button>
-            </div>
-        </div>
         </div>
     );
-};
+}
